Fix for Bug 26270 - "R HDF layer still emitting terrifying warning messages"
If you provide a path like "/a/b" to H5Lexists, but "/a" doesn't
exist, it will fail, giving a hideous error message.  The failure is
actually benign (since /a/b definitely doesn't exist) but error
messages are absolutely terrifying to end users.

From the HDF5 API spec: (http://www.hdfgroup.org/HDF5/doc/RM/RM_H5L.html#Link-Exists)

"Note that H5Lexists verifies only that the target link exists. If
name includes either a relative path or an absloute path to the target
link, intermediate steps along the path must be verified before the
existence of the target link can be safely checked. If the path is not
verified and an intermediate element of the path does not exist,
H5Lexists will fail."